// projects/ngx-table/src/lib/types/data-table-translations.type.ts
export interface DataTableTranslations {
  headerReload?: string;
  headerColumnSelector?: string;
  headerColumnSelectorAdded?: string;
  headerColumnSelectorRemoved?: string;
  indexColumn?: string;
  selectColumn?: string;
  selectRow?: string;
  selectAllRows?: string;
  expandColumn?: string;
  expandRow?: string;
  sortedAscending?: string;
  sortedDescending?: string;
  paginationLimit?: string;
  paginationText?: string;
  paginationTotalPages?: string;
  firstPage?: string;
  prevPage?: string;
  pageNumberLabel?: string;
  pageNumber?: string;
  pageNumberNofM?: string;
  nextPage?: string;
  lastPage?: string;
  loadingText?: string;
  loadedText?: string;
}